+++ SQLiteAndRecyclerViewZadanie1/app/src/main/java/com/example/sqliteandrecyclerviewzadanie1/Note.java
@@ -1,14 +1,14 @@
 package com.example.sqliteandrecyclerviewzadanie1;
 
 public class Note {
-    private int id;
+    private long id;
     private String text;
 
-    public Note(int id, String text) {
+    public Note(long id, String text) {
         this.id = id;
         this.text = text;
     }
 
-    public int getId() { return id; }
+    public long getId() { return id; }
     public String getText() { return text; }
 }
